requirements.txt
assemblyai==0.59.0
python-dotenv==1.2.2
